Add basic animation.
Can now produce animated files of a fractal zoom.

I've noted that the zoom feels like it "speeds up" as it
progresses, a consequence of the linear frame splits when
parameterizing the configurations. To fix that, we'll have
to account for the window size somehow and use more of
our frame budget for the smaller windows.